+++ work1/lad-1117/Campus_IF_Love/Campus_IF_Love.py
@@ -1,0 +1,6 @@
+from Game import Game
+
+if __name__ == "__main__":
+    game = Game()
+    game.start()
+
