--- work1/lad-1117/Character.py
@@ -1,104 +1,0 @@
-class Character:
-    def __init__(self, name, role, affinity=0):
-        self.name = name
-        self.role = role
-        self.affinity = affinity
-
-
-
-    def talk(self):
-        print(f"你正在和{self.name}对话...")
-        all_dias = DIALOGUES.get(self.name)
-        affinity_dias = self.affinity//20 #选择哪一级对话
-        new_dias = all_dias[affinity_dias]
-        sys_dias = new_dias.get('text')
-        option_a = new_dias.get('optionA')
-        option_b = new_dias.get('optionB')
-        print(f'{self.name}:{sys_dias}\n1:{option_a}\n2:{option_b}')
-        while True:
-            your_choice = input('输入你的选择:')
-            if your_choice == "1":
-                self.change_affinity(5) #好感度随正确选项上升
-                break
-            elif your_choice == "2":
-                self.change_affinity(-(affinity_dias+1)*2) #好感度随错误选项下降
-                break
-            else:
-                print('请重新输入')
-            continue
-
-
-    def give_gift(self, gift):
-        print(f"你送给 {self.name} 一份 {gift}。")
-        now_gift = GIFT_EFFECTS.get(gift)
-
-        if now_gift == {"default": -10}:
-            gift_affinity = -10
-        else:
-            gift_affinity = now_gift[self.name]
-
-        if gift_affinity >= 0:
-            print(f'你送出的礼物给 {self.name} 增加了 {gift_affinity} 的好感度。')
-        else:
-            print(f'你送出的礼物给 {self.name} 减少了 {-gift_affinity} 的好感度。')
-
-        self.change_affinity(gift_affinity)
-        pass
-
-
-
-    def change_affinity(self, value):
-        self.affinity += value
-        print(f"{self.name} 的好感度变化 {value} -> 当前好感度：{self.affinity}")
-
-
-
-    def check_affinity(self):
-        print(f'当前 {self.name} 的好感度是 {self.affinity} 。')
-
-
-
-    def check_ending(self):
-        if self.affinity >= 100:
-            print(f"恭喜！你和 {self.name} 的故事进入了结局线！")
-            return True
-        return False
-
-
-
-
-DIALOGUES = {
-    "学姐": [
-        {"text": "你是新来的吧？画画过吗？", "optionA": "其实我更擅长写代码，但我愿意试试画画。", "optionB": "画画？那种小孩子的东西我可没兴趣。"},
-        {"text": "这幅画……构图还算不错，你是随便画的吗？", "optionA": "嗯，我是随手画的，没想到能得到你的认可。", "optionB": "不，我可是很认真的！因为你在看嘛。"},
-        {"text": "加入社团后，不能三天打鱼两天晒网。", "optionA": "放心吧，我会认真对待。", "optionB": "社团只是玩玩而已吧？不要太严肃了。"},
-        {"text": "我平时比较严格，你会不会觉得我很不好相处？", "optionA": "严格才好啊，我喜欢有目标的人。", "optionB": "嗯……确实有点难相处。"},
-        {"text": "今天的社团活动结束了，你要不要和我一起收拾？", "optionA": "当然可以，我来帮你。", "optionB": "算了吧，我还有事要做。"}
-    ],
-    "小白": [
-        {"text": "Python 里的 for 循环，我总是写不对……", "optionA": "来，我教你一个小技巧！", "optionB": "这种简单的东西，你都不会？"},
-        {"text": "你看我借了好多书，会不会显得很傻？", "optionA": "不会啊，这说明你很爱学习，很可爱。", "optionB": "嗯……的确有点太贪心了。"},
-        {"text": "写代码的时候，我总喜欢喝奶茶……你呢？", "optionA": "我也是！来，下次我请你一杯。", "optionB": "我只喝水，健康第一。"},
-        {"text": "你会不会觉得我太依赖你了？", "optionA": "依赖我也没关系，我喜欢这种感觉。", "optionB": "嗯……是有点吧。"},
-        {"text": "要不要一起留在图书馆自习？我想有人陪。", "optionA": "好啊，我正好也要复习。", "optionB": "算了，我还是回宿舍打游戏吧。"}
-    ],
-    "姐姐": [
-        {"text": "你也喜欢在咖啡店看书吗？", "optionA": "是啊，这里很安静，很适合思考。", "optionB": "其实我只是随便找个地方坐。"},
-        {"text": "研究生的生活，其实没有你想象的那么光鲜。", "optionA": "我能理解，压力一定很大吧。", "optionB": "那我还是不要考研了。"},
-        {"text": "你觉得，什么样的人最值得依靠？", "optionA": "稳重冷静的人。", "optionB": "长得好看的人。"},
-        {"text": "我常常一个人坐到很晚，你不会觉得孤单吗？", "optionA": "有时候孤单是种享受。", "optionB": "一个人太寂寞了，我受不了。"},
-        {"text": "如果有一天，我遇到困难，你会帮我吗？", "optionA": "当然会，你不用一个人扛着。", "optionB": "我大概帮不了你吧。"}
-    ]
-}
-
-
-GIFT_EFFECTS = {
-    # 通用 / 默认 值
-    "鲜花": {"学姐": 10, "小白": 10, "姐姐": 15},
-    "编程笔记": {"学姐": 5, "小白": 15, "姐姐": 15},
-    "奶茶": {"学姐": 20, "小白": 20, "姐姐": 20},
-    "奇怪的石头": {"default": -10},  # 所有人 -10
-    "精致的钢笔": {"学姐": 20, "小白": 10, "姐姐": 20},
-    "可爱玩偶": {"学姐": 10, "小白": 20, "姐姐": 10},
-    "夜宵外卖": {"学姐": 0, "小白": 5, "姐姐": -5}
-}
